revision: rename wasm file
